Refactor colour calibration code. This is a bit experimental, being the first real C++ programming I've done in about a decade.
- Use std::array? Probably better than the default. *shrug*
	- Remove preprocessor-based min and max from types.h. Can't #include <array> with those guys around.
- Use c++17 mode, in a failed attempt to get non-double-bracketed array initializers and std::clamp.
- Split Camera::setCCMatrix into two, one function to calculate the ultimate matrix and one to set the matrix correctly.
- Give long, sensible names to a bunch of stuff relating to the calculations.
	- Things make a lot more sense when named, it seems.

+++ src/videoRecord.h
@@ -95,9 +95,9 @@ root@dm814x-evm:~#
 
  **/
 
-#define ENCODER_MIN_H_RES		96	//Minimum resoluiton the TI encoder can run at
-#define ENCODER_MIN_V_RES		80
-#define ENCODER_H_RES_INC		16	//Horizontal resolution must be a multiple of 16
+#define ENCODER_MIN_H_RES		((UInt32) 96)	//Minimum resoluiton the TI encoder can run at
+#define ENCODER_MIN_V_RES		((UInt32) 80)
+#define ENCODER_H_RES_INC		((UInt32) 16)	//Horizontal resolution must be a multiple of 16
 /*
 typedef enum VideoRecordErrortype
 {
